backend/analytics/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlmodel
psycopg[binary]
